Backend/requirements.txt
# backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
transformers==4.36.0
torch==2.1.0
sentence-transformers==2.2.2
pinecone-client==3.0.1
python-dotenv==1.0.0
pydantic==2.5.0
langchain==0.0.340
langchain-community==0.0.10
streamlit==1.28.0
streamlit-chat==0.1.1
python-multipart==0.0.6
httpx==0.25.1
ollama==0.1.6
requests==2.31.0
backoff==2.2.1
huggingface-hub==0.20.0